Unity Cookbook 3D
This repository contains minimal codes that run very basic functions in
Unity3D. When I was watching/reading tutorials and Q&As on the
internet, I encounter issues getting some of people describing to work
on my system. Not sure if it’s due to different system configuration,
Unity version, operating system, etc.

Hopefully, this repo will help people wants to learn Unity. Feel free
to branch, contribute, or use the code as you pleased.

The code is written using Unity 5.5 for Mac.

To install:
1. Download the whole repository as a zip and extract. You can also do
‘git clone’
2. Launch Unity and create a new 3D project.
3. Drag the whole repository folder into Unity asset folder.